requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic[email]
